backend/requirements.txt
flask
flask-cors
Pillow
schedule
tqdm
pdfplumber
opencv-python-headless
ubelt
surya-ocr
easyocr